+++ tests/Equinox.CosmosStore.Integration/CosmosCoreIntegration.fs
@@ -1,7 +1,7 @@
-﻿module Equinox.Cosmos.Integration.CoreIntegration
+﻿module Equinox.CosmosStore.Integration.CoreIntegration
 
-open Equinox.Cosmos.Core
-open Equinox.Cosmos.Integration.Infrastructure
+open Equinox.CosmosStore.Core
+open Equinox.CosmosStore.Integration.Infrastructure
 open FsCodec
 open FSharp.Control
 open Newtonsoft.Json.Linq
@@ -93,14 +93,14 @@ type Tests(testOutputHelper) =
 
     let verifyCorrectEventsEx direction baseIndex (expected: IEventData<_>[]) (xs: ITimelineEvent<byte[]>[]) =
         let xs, baseIndex =
-            if direction = Equinox.Cosmos.Store.Direction.Forward then xs, baseIndex
+            if direction = Equinox.CosmosStore.Store.Direction.Forward then xs, baseIndex
             else Array.rev xs, baseIndex - int64 (Array.length expected) + 1L
         test <@ [for i in 0..expected.Length - 1 -> baseIndex + int64 i] = [for r in xs -> r.Index] @>
         test <@ [for e in expected -> e.EventType] = [ for r in xs -> r.EventType ] @>
         for i,x,y in Seq.mapi2 (fun i x y -> i,x,y) [for e in expected -> e.Data] [for r in xs -> r.Data] do
             verifyUtf8JsonEquals i x y
-    let verifyCorrectEventsBackward = verifyCorrectEventsEx Equinox.Cosmos.Store.Direction.Backward
-    let verifyCorrectEvents = verifyCorrectEventsEx Equinox.Cosmos.Store.Direction.Forward
+    let verifyCorrectEventsBackward = verifyCorrectEventsEx Equinox.CosmosStore.Store.Direction.Backward
+    let verifyCorrectEvents = verifyCorrectEventsEx Equinox.CosmosStore.Store.Direction.Forward
 
     [<AutoData(SkipIfRequestedViaEnvironmentVariable="EQUINOX_INTEGRATION_SKIP_COSMOS")>]
     let ``appendAtEnd and getNextIndex`` (extras, TestStream streamName) = Async.RunSynchronously <| async {
@@ -254,7 +254,7 @@ type Tests(testOutputHelper) =
         verifyCorrectEvents 0L expected res
         test <@ [EqxAct.ResponseForward; EqxAct.QueryForward] = capture.ExternalCalls @>
         let queryRoundTripsAndItemCounts = function
-            | EqxEvent (Equinox.Cosmos.Store.Log.Event.Query (Equinox.Cosmos.Store.Direction.Forward, responses, { count = c })) -> Some (responses,c)
+            | EqxEvent (Equinox.CosmosStore.Store.Log.Event.Query (Equinox.CosmosStore.Store.Direction.Forward, responses, { count = c })) -> Some (responses,c)
             | _ -> None
         // validate that, despite only requesting max 1 item, we only needed one trip (which contained only one item)
         [1,1] =! capture.ChooseCalls queryRoundTripsAndItemCounts
@@ -320,7 +320,7 @@ type Tests(testOutputHelper) =
         test <@ [EqxAct.ResponseBackward; EqxAct.QueryBackward] = capture.ExternalCalls @>
         // validate that, despite only requesting max 1 item, we only needed one trip, bearing 5 items (from which one item was omitted)
         let queryRoundTripsAndItemCounts = function
-            | EqxEvent (Equinox.Cosmos.Store.Log.Event.Query (Equinox.Cosmos.Store.Direction.Backward, responses, { count = c })) -> Some (responses,c)
+            | EqxEvent (Equinox.CosmosStore.Store.Log.Event.Query (Equinox.CosmosStore.Store.Direction.Backward, responses, { count = c })) -> Some (responses,c)
             | _ -> None
         [1,5] =! capture.ChooseCalls queryRoundTripsAndItemCounts
         verifyRequestChargesMax 4 // 3.24 // WAS 3 // 2.98